rename: pages 폴더구조 변경
모의면접과 관련된 컴포넌트는 room 폴더 안으로 옮김.
